PerformanceImpl.cpp: size_t
svn:sync-xref-src-repo-rev: r45326

+++ src/VBox/Main/PerformanceImpl.cpp
@@ -1,4 +1,4 @@
-/* $Id: PerformanceImpl.cpp 18252 2009-03-25 12:59:12Z noreply@oracle.com $ */
+/* $Id: PerformanceImpl.cpp 18487 2009-03-29 01:35:27Z knut.osmundsen@oracle.com $ */
 
 /** @file
  *
@@ -438,7 +438,7 @@ PerformanceCollector::QueryMetricsData (ComSafeArrayIn (IN_BSTR, metricNames),
         retScales[i] = (*it)->getScale();
         retSequenceNumbers[i] = sequenceNumber;
         retLengths[i] = length;
-        retIndices[i] = flatIndex;
+        retIndices[i] = (ULONG)flatIndex;
         flatIndex += length;
     }
